refactor: lift a cohesive unit out of seven files at the max-lines limit
Download naming, the browse filter model, the print response shape, the
community client's type guards, the layout export file helpers, the
supporters page flourishes and the divider-tilt list each move to a sibling
module; the originals re-export the public names.
