+++ api/community/prints.ts
@@ -60,6 +60,8 @@ import {
   writeCommunityPrint,
 } from '../lib/communityPrintStore.js';
 import type { CommunityPrintRecord, CommunityPrintStatus } from '../lib/communityPrintStore.js';
+import { toPrintResponse } from '../lib/communityPrintResponse.js';
+import type { PrintResponse } from '../lib/communityPrintResponse.js';
 
 const COMMUNITY_DESIGN_ID_REGEX = /^[a-zA-Z0-9]{12}$/;
 const AUTHOR_PUBLIC_ID_REGEX = /^[a-f0-9]{32}$/;
@@ -138,56 +140,6 @@ async function readOptionalSession(req: VercelRequest): Promise<SessionRecord | 
   } catch {
     return null;
   }
-}
-
-interface PrintResponse {
-  id: string;
-  designId: string;
-  authorPublicId: string;
-  authorName: string;
-  photos: string[];
-  photoThumbs: string[];
-  /** Every field omitted rather than nulled when unreported, so a client that
-      reads `settings.nozzleMm` gets `undefined` and cannot format a 0. */
-  settings: {
-    material?: string;
-    nozzleMm?: number;
-    layerHeightMm?: number;
-    printMinutes?: number;
-    filamentGrams?: number;
-    printer?: string;
-    printerOther?: string;
-  };
-  fitVerdict: string;
-  note: string;
-  createdAt: number;
-  updatedAt: number;
-  status: string;
-}
-
-function toPrintResponse(record: CommunityPrintRecord): PrintResponse {
-  return {
-    id: `${record.designId}:${record.authorPublicId}`,
-    designId: record.designId,
-    authorPublicId: record.authorPublicId,
-    authorName: record.authorName,
-    photos: record.photos,
-    photoThumbs: record.photoThumbs,
-    settings: {
-      ...(record.material !== null && { material: record.material }),
-      ...(record.nozzleMm !== null && { nozzleMm: record.nozzleMm }),
-      ...(record.layerHeightMm !== null && { layerHeightMm: record.layerHeightMm }),
-      ...(record.printMinutes !== null && { printMinutes: record.printMinutes }),
-      ...(record.filamentGrams !== null && { filamentGrams: record.filamentGrams }),
-      ...(record.printer !== null && { printer: record.printer }),
-      ...(record.printerOther !== '' && { printerOther: record.printerOther }),
-    },
-    fitVerdict: record.fitVerdict,
-    note: record.note,
-    createdAt: record.createdAt,
-    updatedAt: record.updatedAt,
-    status: record.status,
-  };
 }
 
 async function handleList(
