+++ src/features/community/api/client.ts
@@ -1,3 +1,13 @@
+import {
+  isDesignResponse,
+  isDetailResponse,
+  isListPage,
+  isCapabilities,
+  isOptionalNumber,
+  parseHiddenReason,
+} from './clientGuards';
+import type { CommunityListPage, CommunityCapabilities } from './clientGuards';
+export type { CommunityCapabilities } from './clientGuards';
 import type { Result } from '@/core/result';
 import { ok, err, isErr } from '@/core/result';
 import { isApiErrorResponse } from '@/core/api/mapApiError';
@@ -14,14 +24,10 @@ import type {
   CommunityHiddenReason,
   CommunityReportReason,
 } from '@/shared/types/community';
-import type { CommunityFeatureReason } from '@/shared/types/community';
 import {
-  COMMUNITY_CATEGORIES,
-  COMMUNITY_FEATURE_REASONS,
   COMMUNITY_REPORT_NOTE_MAX_LENGTH,
   COMMUNITY_REPORT_REASONS,
 } from '@/shared/types/community';
-import { TECHNIQUE_CONFIG } from '@/shared/types/exampleTechniques';
 import { isRecord } from '@/shared/utils/isRecord';
 
 const COMMUNITY_ENDPOINT = '/api/community';
@@ -103,33 +109,6 @@ function isPublishResult(value: unknown): value is CommunityPublishResult {
   return isRecord(value) && typeof value.id === 'string' && typeof value.url === 'string';
 }
 
-function isCommunityDesign(value: unknown): value is CommunityDesign {
-  if (!isRecord(value)) return false;
-  return (
-    typeof value.id === 'string' &&
-    typeof value.authorPublicId === 'string' &&
-    typeof value.authorName === 'string' &&
-    typeof value.name === 'string' &&
-    typeof value.description === 'string' &&
-    isKnownCategory(value.category) &&
-    Array.isArray(value.techniques) &&
-    value.techniques.every(isKnownTechnique) &&
-    (isRecord(value.params) ||
-      (value.kind === 'assembly' && isRecord(value.envelope) && isRecord(value.structure))) &&
-    isRecord(value.metrics) &&
-    (value.lineage === null || isRecord(value.lineage)) &&
-    Array.isArray(value.thumbnails) &&
-    typeof value.meshUrl === 'string' &&
-    typeof value.createdAt === 'number' &&
-    typeof value.updatedAt === 'number' &&
-    (value.status === 'live' || value.status === 'hidden' || value.status === 'removed')
-  );
-}
-
-function isDesignResponse(value: unknown): value is { design: CommunityDesign } {
-  return isRecord(value) && isCommunityDesign(value.design);
-}
-
 export interface CommunityDesignDetail {
   design: CommunityDesign;
   /** Server-verified against the session's published set, never a client-sent id. */
@@ -153,10 +132,6 @@ export interface CommunityDesignDetail {
   hiddenReasonCategory: CommunityReportReason | null;
 }
 
-function isOptionalNumber(value: unknown): boolean {
-  return value === undefined || typeof value === 'number';
-}
-
 function isCountsShape(value: unknown): value is CommunityDesignCounts {
   return (
     isRecord(value) &&
@@ -168,10 +143,6 @@ function isCountsShape(value: unknown): value is CommunityDesignCounts {
   );
 }
 
-function parseHiddenReason(value: unknown): CommunityHiddenReason | null {
-  return value === 'reports' || value === 'denylist' || value === 'moderation' ? value : null;
-}
-
 function parseReportReason(value: unknown): CommunityReportReason | null {
   return typeof value === 'string' &&
     (COMMUNITY_REPORT_REASONS as readonly string[]).includes(value)
@@ -179,112 +150,9 @@ function parseReportReason(value: unknown): CommunityReportReason | null {
     : null;
 }
 
-function isDetailResponse(value: unknown): value is {
-  design: CommunityDesign;
-  isOwner?: boolean;
-  counts?: unknown;
-  likedByMe?: unknown;
-  authorIsSupporter?: unknown;
-  hiddenReason?: unknown;
-  hiddenReasonCategory?: unknown;
-} {
-  return isDesignResponse(value) && (!('isOwner' in value) || typeof value.isOwner === 'boolean');
-}
-
-const KNOWN_TECHNIQUES: readonly string[] = Object.keys(TECHNIQUE_CONFIG);
-
-function isKnownCategory(value: unknown): value is CommunityCategory {
-  return typeof value === 'string' && (COMMUNITY_CATEGORIES as readonly string[]).includes(value);
-}
-
-function isKnownTechnique(value: unknown): boolean {
-  return typeof value === 'string' && KNOWN_TECHNIQUES.includes(value);
-}
-
-function isKnownFeatureReason(value: unknown): value is CommunityFeatureReason {
-  return (
-    typeof value === 'string' && (COMMUNITY_FEATURE_REASONS as readonly string[]).includes(value)
-  );
-}
-
-function isCommunityCard(value: unknown): value is CommunityCard {
-  if (!isRecord(value)) return false;
-  const counts: unknown = value.counts;
-  const metrics: unknown = value.metrics;
-  return (
-    typeof value.id === 'string' &&
-    typeof value.name === 'string' &&
-    typeof value.authorName === 'string' &&
-    typeof value.authorPublicId === 'string' &&
-    isKnownCategory(value.category) &&
-    Array.isArray(value.techniques) &&
-    value.techniques.every(isKnownTechnique) &&
-    isRecord(metrics) &&
-    typeof metrics.width === 'number' &&
-    typeof metrics.depth === 'number' &&
-    typeof metrics.height === 'number' &&
-    typeof metrics.gridUnitMm === 'number' &&
-    typeof value.thumbnailUrl === 'string' &&
-    typeof value.isRemix === 'boolean' &&
-    (value.parentId === undefined || typeof value.parentId === 'string') &&
-    typeof value.featured === 'boolean' &&
-    // A reason outside the union fails the whole card rather than flowing
-    // through as a typed value the UI will index into a label map.
-    (value.featureReason === undefined || isKnownFeatureReason(value.featureReason)) &&
-    isRecord(counts) &&
-    typeof counts.likes === 'number' &&
-    typeof counts.remixes === 'number' &&
-    typeof counts.exports === 'number' &&
-    isOptionalNumber(counts.opens) &&
-    isOptionalNumber(counts.views) &&
-    typeof value.createdAt === 'number' &&
-    typeof value.updatedAt === 'number' &&
-    (value.status === 'live' || value.status === 'hidden' || value.status === 'removed') &&
-    (value.hiddenReason === undefined || parseHiddenReason(value.hiddenReason) !== null)
-  );
-}
-
-interface CommunityListPage {
-  items: CommunityCard[];
-  nextCursor: string | null;
-  /** Ids on this page the session user has liked; empty for anonymous callers. */
-  likedIds?: string[];
-  /**
-   * Author public ids on this page whose supporter badge is public. Keyed by
-   * author rather than by design, so one publisher with several cards on a
-   * page costs one entry. Optional: an older deployment omits it and nothing
-   * is badged.
-   */
-  supporterAuthorIds?: string[];
-  /**
-   * Slots in the server's index, so the whole thing can be requested as
-   * concurrent windows. Optional: an older deployment omits it and the fetch
-   * falls back to paging sequentially.
-   */
-  indexSlots?: unknown;
-}
-
 /** Seed the per-card supporter flag from a page's author-keyed sidecar. */
 function supporterAuthorSet(page: CommunityListPage): Set<string> {
   return new Set(page.supporterAuthorIds ?? []);
-}
-
-function isListPage(value: unknown): value is CommunityListPage {
-  return (
-    isRecord(value) &&
-    Array.isArray(value.items) &&
-    value.items.every(isCommunityCard) &&
-    (value.nextCursor === null || typeof value.nextCursor === 'string') &&
-    (value.likedIds === undefined ||
-      (Array.isArray(value.likedIds) && value.likedIds.every((id) => typeof id === 'string'))) &&
-    (value.supporterAuthorIds === undefined ||
-      (Array.isArray(value.supporterAuthorIds) &&
-        value.supporterAuthorIds.every((id) => typeof id === 'string'))) &&
-    // indexSlots is deliberately not validated here. It is an optimisation
-    // hint, and a corrupt hint must not reject a page of real designs; it is
-    // read through readIndexSlots, which treats anything unusable as absent.
-    true
-  );
 }
 
 function isDeleteResponse(value: unknown): value is { success: true } {
@@ -296,26 +164,6 @@ export async function communityFetch(input: string, init: RequestInit): Promise<
   // forced sign-out event would clear the sync outbox and flip every tab
   // anonymous.
   return apiFetch(input, { ...init, suppressForcedSignOut: true });
-}
-
-/**
- * The server switches the client cannot infer. `community_showcase` is a
- * per-user Labs flag over the UI; these are deployment kill switches, and
- * nothing local reflects them.
- */
-export interface CommunityCapabilities {
-  publishEnabled: boolean;
-  printsEnabled: boolean;
-  requireDescription: boolean;
-}
-
-function isCapabilities(value: unknown): value is CommunityCapabilities {
-  return (
-    isRecord(value) &&
-    typeof value.publishEnabled === 'boolean' &&
-    typeof value.printsEnabled === 'boolean' &&
-    typeof value.requireDescription === 'boolean'
-  );
 }
 
 /**
